requirements.txt
Adafruit-GPIO>=0.6.5
adafruit-pca9685
paho-mqtt
